requirements.txt
Django==1.8.6
django-appconf==1.0.1
django-ckeditor==5.0.2
django-compressor==1.5
six==1.10.0
Wand==0.4.1
wheel==0.24.0
